Renamed getMobsToHeal so it doesn't get caught by the kotlin auto-getter

+++ src/main/kotlin/com/aaronhowser1/pitchperfect/utils/ServerUtils.kt
@@ -15,7 +15,7 @@ object ServerUtils {
         )
     }
 
-    fun getNearbyLivingEntities(livingEntity: LivingEntity, range: Int): List<LivingEntity> {
+    fun getNearbyLivingEntities(livingEntity: LivingEntity, range: Float): List<LivingEntity> {
         return livingEntity.getLevel().getEntities(
             livingEntity,
             AABB(
